Added Master Manager
Created a singleton type for scriptable objects, of which MasterManager belongs. This manager holds onto the game settings and when the TestConnect script runs, it sets up the settings for the connection based on the MasterManager's reference to the GameSettings scriptable object (non-singleton because you may want to have additional setttings, such as different profiles or accounts).